configured jwt to create token while creating session and authenticate user while deleting a post

+++ controllers/api/v1/posts_api.js
@@ -24,19 +24,23 @@ module.exports.index = async function(req, res){
 module.exports.destroy = async function(req, res){
     try{
         let post = await Post.findById(req.params.id);
+        if(post.user == req.user.id){
+            post.remove();
 
-        post.remove();
-
-        await Comment.deleteMany({post : req.params.id});
-
-
-        return res.status(200).json({
-            message : "Post Deleted"
-        })
+            await Comment.deleteMany({post : req.params.id});
+            
+            return res.status(200).json({
+                message : "Post Deleted"
+            });
+        }else{
+            return res.status(401).json({
+                message : "You cannot delete this post!"
+            });
+        }
 
     }catch(err){
         return res.status(500).json({
             message : "Internal Server Error"
-        })
+        });
     };
 };
